% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlefont.R
\docType{data}
\name{effects}
\alias{effects}
\title{font effects}
\format{data frame with columns
\describe{
\item{name}{readable name of the effect}
\item{effect}{name of the effect}
\item{class}{css class}
}}
\source{
\url{https://developers.google.com/fonts/docs/getting_started}
}
\usage{
effects
}
\description{
font effects
}
\keyword{datasets}

