% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlefont.R
\name{fonts}
\alias{fonts}
\title{Retrieves the list of fonts}
\usage{
fonts(token = getOption("googlefonts_token"))
}
\arguments{
\item{token}{a google font api token.}
}
\value{
a data frame with columns
\describe{
  \item{family}{the font family}
  \item{files}{}
}
}
\description{
Retrieves the list of fonts
}
\details{
the function is \code{\link[memoise]{memoise}}d to limit the number of
requests to google per session.
}

