% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlefont.R
\name{googlefont}
\alias{googlefont}
\title{get google font}
\usage{
googlefont(family = "Inconsolata", style = NULL, subset = NULL,
  text = NULL, effect = NULL)
}
\arguments{
\item{family}{font family}

\item{style}{style, e.g. \samp{bold}, \samp{italic}, ...}

\item{subset}{subset, e.g. \samp{cyrillic}, ...}

\item{text}{text}

\item{effect}{font effect}
}
\description{
get google font
}
\references{
Google Fonts. \url{https://fonts.google.com}
}

