% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlefont.R
\name{has_font}
\alias{has_font}
\title{check if the font is available}
\usage{
has_font(family, variant, subset)
}
\arguments{
\item{family}{font family}

\item{variant}{font variant, .e.g \code{"italic"}}

\item{subset}{font subset, e.g. \code{"latin"}}
}
\description{
check if the font is available
}
\examples{
\dontrun{
  has_font( "Inconsolata")
}
}

