% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlefont.R
\name{use_google_font}
\alias{use_google_font}
\title{get google font}
\usage{
use_google_font(family = "Inconsolata", style = NULL, subset = NULL,
  text = NULL, effect = NULL)
}
\arguments{
\item{family}{font family}

\item{style}{style, e.g. \samp{bold}, \samp{italic}, ...}

\item{subset}{subset, e.g. \samp{cyrillic}, ...}

\item{text}{text}

\item{effect}{font effect}
}
\description{
get google font
}
\examples{
\dontrun{
use_google_font( "Rancho", effect = "shadow-multiple")
use_google_font( "Inconsolata" )
}
}
\references{
Google Fonts. \url{https://fonts.google.com}
}

